Fix weightingFactor, was wrong data type

+++ src/Configuration.cpp
@@ -7,12 +7,12 @@
 
 #include "Configuration.h"
 
-unsigned int Configuration::sleepTime           = 10;
-unsigned int Configuration::weightingFactor     = .5;
-float Configuration::energyPerSamplingCycle     = .0002;
-float Configuration::energyPerStorageCycle      = .00006;
-const float Configuration::energyStorageEmpty   = 1.0;
-const float Configuration::energyStorageFull    = 2.5;
+unsigned int   Configuration::sleepTime       = 10;
+float Configuration::weightingFactor          = .5;
+float Configuration::energyPerSamplingCycle   = .0002;
+float Configuration::energyPerStorageCycle    = .00006;
+const float Configuration::energyStorageEmpty = 1.0;
+const float Configuration::energyStorageFull  = 2.5;
 
 void Configuration::updateConfiguration( uint8_t *configPacket )
 {
